bootstrap/bootstrap.go(checkoutPlugin): Optionally clone fresh.
If the user has specified `BUILDKITE_PLUGINS_FORCE_PULL=true`, we check whether the plugin already
exists and delete it, which will cause the subsequent code to (as per usual, as if from a blank
slate) make a fresh clone of the plugin repo.

This is very useful during plugin development, so that the plugin developer can simply point at a
branch of their plugin and be certain that they're always getting the latest version when trying it
out.

I went in a few circles thinking about "better" ways of accomplishing this, specifically by trying
to fetch & checkout instead of just blowing away the entire local copy.  But because Git's
pull/fetch/merge/reset/checkout behaviour is _very_ subtle (e.g., i need this feature to work if the
branch has been force-pushed, if it's actually a tag, if it's a simple fast-forward merge, etc.) i
figured it's probably most robust to just rely on the battle-tested fresh-clone code that's already
there.  Not to mention that there is SSH known-hosts magic that may or may not cause trouble when
executing a `git fetch`, i thought it's probably way overkill to try and be smart.  Plus, how large
are plugin repositories really going to be in the end, and if this is a dev-only feature, what's the
likelihood of getting rate limited?